Fix logic for overriden templates
