+++ examples/DemoApp/DemoApp.Contracts/Templates/GuidId.liquid
@@ -5,6 +5,8 @@ using System.Runtime.InteropServices.ComTypes;
 using System.Text.Json.Serialization;
 using NForza.Cyrus.Abstractions;
 
+// Overridden template
+
 namespace {{ Namespace }};
 
 [JsonConverter(typeof({{ Name }}JsonConverter))]
